Move Streams/Serial.hs to Internal/Data/Stream
