+++ src/Streamly.hs
@@ -228,7 +228,7 @@ import Streamly.Internal.Data.Stream.Ahead
 import Streamly.Internal.Data.Stream.Async
 import Streamly.Internal.Data.Stream.Combinators
 import Streamly.Streams.Parallel
-import Streamly.Streams.Serial
+import Streamly.Internal.Data.Stream.Serial
 import Streamly.Internal.Data.Stream.StreamK hiding (serial)
 import Streamly.Internal.Data.Stream.Zip
