+++ src/Streamly/FileSystem/FD.hs
@@ -137,7 +137,7 @@ import qualified GHC.IO.FD as FD
 import qualified GHC.IO.Device as RawIO
 
 import Streamly.Internal.Memory.Array.Types (Array(..), byteLength, defaultChunkSize)
-import Streamly.Streams.Serial (SerialT)
+import Streamly.Internal.Data.Stream.Serial (SerialT)
 import Streamly.Internal.Data.Stream.StreamK.Type (IsStream, mkStream)
 
 #if !defined(mingw32_HOST_OS)
