+++ src/Streamly/Internal/Data/Stream/Ahead.hs
@@ -47,7 +47,7 @@ import GHC.Exts (inline)
 import qualified Data.Heap as H
 
 import Streamly.Internal.Data.Stream.SVar (fromSVar)
-import Streamly.Streams.Serial (map)
+import Streamly.Internal.Data.Stream.Serial (map)
 import Streamly.Internal.Data.SVar
 import Streamly.Internal.Data.Stream.StreamK
        (IsStream(..), Stream, mkStream, foldStream, foldStreamShared,
