+++ src/Streamly/Internal/Data/Stream/Async.hs
@@ -58,7 +58,7 @@ import qualified Data.Set as S
 
 import Streamly.Internal.Data.Atomics (atomicModifyIORefCAS)
 import Streamly.Internal.Data.Stream.SVar (fromSVar)
-import Streamly.Streams.Serial (map)
+import Streamly.Internal.Data.Stream.Serial (map)
 import Streamly.Internal.Data.SVar
 import Streamly.Internal.Data.Stream.StreamK
        (IsStream(..), Stream, mkStream, foldStream, adapt, foldStreamShared,
