+++ src/Streamly/Internal/Data/Stream/Combinators.hs
@@ -31,7 +31,7 @@ import Data.Int (Int64)
 
 import Streamly.Internal.Data.SVar
 import Streamly.Internal.Data.Stream.StreamK
-import Streamly.Streams.Serial (SerialT)
+import Streamly.Internal.Data.Stream.Serial (SerialT)
 
 -------------------------------------------------------------------------------
 -- Concurrency control
