+++ src/Streamly/Internal/Memory/ArrayStream.hs
@@ -50,7 +50,7 @@ import Foreign.Storable (Storable(..))
 import Prelude hiding (length, null, last, map, (!!), read, concat)
 
 import Streamly.Internal.Memory.Array.Types (Array(..), length)
-import Streamly.Streams.Serial (SerialT)
+import Streamly.Internal.Data.Stream.Serial (SerialT)
 import Streamly.Internal.Data.Stream.StreamK.Type (IsStream)
 
 import qualified Streamly.Internal.Memory.Array as A
